+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - `get_similar`
     - `genres`, `genre`, `genre_artists`, `genre_albums`, `genre_songs`
     - `bookmarks`, `get_bookmark`, `bookmark_create`, `bookmark_edit`, `bookmark_delete`
+    - `live_streams`, `live_stream`
   * Type `album_artist` to the method `get_indexes` (but without support for the args `filter`, `add`, `update`)
   * Fields `time`, `albumcount`, and `songcount` to the `artist` and `album` type results
   * Field `disk` to `song` type results